テーブル構成の修正
